Módulo auth (adaptação Flask-Login) - Views e métodos relacionados com login agora foram colocados num módulo 'auth', já registrado no app
